backend/requirements-minimal.txt
# Minimal requirements for Python 3.14 (no compilation needed)
fastapi
uvicorn
aiogram
google-generativeai
beautifulsoup4
httpx
requests
python-dotenv
loguru
